Bugfix: Weird grid-mass interactions on pause
Bug: On pause, the grid-mass interactions in certain scenes seem to invert (i.e. grid interacts with mass only if it is outside the grid)
Cause: "Double-adding" of same mass object, which disables grid + causes weird behavior
Fix: Added check to prevent adding of same mass object to List
