--- POINT-VR-Chapter-1/Assets/POINT/Feature3/Constrains.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: d00f00d0d8fb470fb68084586be9a558
-timeCreated: 1670310199
